some refactoring, add unit tests for ConcretizeLike
